requirements.txt
annotated-types==0.5.0
anyio==4.0.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
Flask==2.3.3
Flask-Testing==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7